Move Cloudflare error page detection into CloudflareRemoteUITestBase.
